@@ -15,11 +15,11 @@
   },
   "config": {
     "platform": {
-      "php": "8.1"
+      "php": "8.2"
     }
   },
   "require" : {
-    "php": ">=8.1.0",
+    "php": ">=8.2.0",
     "guzzlehttp/guzzle": "^7.5",
     "ext-ctype": "*",
     "ext-json": "*"
