@@ -3,6 +3,7 @@
 declare(strict_types=1);
 
 use Rector\CodeQuality\Rector\Class_\InlineConstructorDefaultToPropertyRector;
+use Rector\CodeQuality\Rector\ClassMethod\InlineArrayReturnAssignRector;
 use Rector\Config\RectorConfig;
 use Rector\Set\ValueObject\LevelSetList;
 use Rector\Set\ValueObject\SetList;
@@ -14,9 +15,16 @@ return static function (RectorConfig $rectorConfig): void {
     ]);
 
     // register a single rule
-     $rectorConfig->rule(InlineConstructorDefaultToPropertyRector::class);
+    $rectorConfig->rule(InlineConstructorDefaultToPropertyRector::class);
+    $rectorConfig->rule(InlineArrayReturnAssignRector::class);
 
-    $rectorConfig->sets([LevelSetList::UP_TO_PHP_81]);
-    $rectorConfig->sets([SetList::CODE_QUALITY]);
-//    $rectorConfig->sets([SetList::CODING_STYLE]);
+
+    $rectorConfig->sets([
+        LevelSetList::UP_TO_PHP_82,
+        SetList::CODE_QUALITY,
+        SetList::CODING_STYLE,
+        SetList::TYPE_DECLARATION,
+//        SetList::DEAD_CODE,
+        SetList::EARLY_RETURN,
+    ]);
 };
