@@ -22,7 +22,14 @@ use stdClass;
  */
 abstract class AbstractWebService
 {
+    /**
+     * @var int
+     */
     final public const LIMIT_MAX = 20;
+
+    /**
+     * @var int
+     */
     final public const TIMEOUT = 30;
 
     protected int $timeout = self::TIMEOUT;
@@ -32,9 +39,6 @@ abstract class AbstractWebService
         'Content-Type' => GuzzleSingleton::APPLICATION_JSON
     ];
 
-    /**
-     * @var Client
-     */
     private Client $client;
 
     /**
@@ -130,7 +134,8 @@ abstract class AbstractWebService
             $options['body'] = $body;
         }
 
-        $guzzleResponse = $msgErr = null;
+        $guzzleResponse = null;
+        $msgErr = null;
         try {
             $guzzleResponse = $this->client->request($method, $endPoint, $options);
         } catch (GuzzleException $guzzleException) {
@@ -143,7 +148,6 @@ abstract class AbstractWebService
 
         throw new WsException($msgErr, 500, null);
     }
-
 
     /**
      * Check response and jsondecode object
@@ -174,17 +178,20 @@ abstract class AbstractWebService
             throw new WsResponseException(WsException::RESP_EMPTY, 500, null);
         }
 
-        if (
-            property_exists($jsonContent, "objects")
-            && property_exists($jsonContent, "meta")
-            && 0 === $jsonContent->meta->total_count
-        ) {
-            throw new WsResponseException(WsException::RESP_EMPTY, 500, null);
+        if (!property_exists($jsonContent, "objects")) {
+            return $contents;
         }
 
-        return $contents;
-    }
+        if (!property_exists($jsonContent, "meta")) {
+            return $contents;
+        }
 
+        if (0 !== $jsonContent->meta->total_count) {
+            return $contents;
+        }
+
+        throw new WsResponseException(WsException::RESP_EMPTY, 500, null);
+    }
 
     /**
      * @throws JsonException
