@@ -8,12 +8,9 @@ namespace AstrobinWs\Response\DTO;
  * Class AstrobinError
  * @package AstrobinWs\Response
  */
-final class AstrobinError implements AstrobinResponse
+final readonly class AstrobinError implements AstrobinResponse
 {
-    /**
-     * AstrobinError constructor.
-     */
-    public function __construct(private readonly ?string $message)
+    public function __construct(private ?string $message)
     {
     }
 
