@@ -16,23 +16,29 @@ use AstrobinWs\Response\DTO\Collection\ListImages;
 final class Collection extends AbstractResponse implements AstrobinResponse
 {
     public int $id;
+
     public string $name;
+
     public ?string $description = null;
 
+
     public string $user;
-    public string|\DateTime $date_created;
-    public string|\DateTime $date_updated;
+
+    public string|null|\DateTime $date_created = null;
+
+    public string|null|\DateTime $date_updated = null;
+
     public ListImages|array|null $images = null;
 
-    public function setDateCreated(string $dateCreated): self
+    public function setDateCreated(?string $dateCreated): self
     {
-        $this->date_created = \DateTime::createFromFormat(QueryFilters::DATE_FORMAT->value, $dateCreated);
+        $this->date_created = ((is_null($dateCreated)) ? null : \DateTime::createFromFormat(QueryFilters::DATE_FORMAT->value, $dateCreated)) ?: null;
         return $this;
     }
 
-    public function setDateUpdated(string $dateUpdated): self
+    public function setDateUpdated(?string $dateUpdated): self
     {
-        $this->date_updated = \DateTime::createFromFormat(QueryFilters::DATE_FORMAT->value, $dateUpdated);
+        $this->date_updated = ((is_null($dateUpdated)) ? null : \DateTime::createFromFormat(QueryFilters::DATE_FORMAT->value, $dateUpdated)) ?: null;
         return $this;
     }
 }
