@@ -14,10 +14,11 @@ use AstrobinWs\Response\DTO\AstrobinResponse;
 final class Today extends AbstractResponse implements AstrobinResponse
 {
     public string $date;
+
     /**
      * Path of image, not Image instance
      */
     public string|Image|null $image = null;
-    public string $resource_uri;
 
+    public string $resource_uri;
 }
