@@ -12,7 +12,6 @@ use AstrobinWs\Response\DTO\Collection\ListImages;
 use AstrobinWs\Response\DTO\Item\Image;
 use DateTime;
 use JsonException;
-use ReflectionException;
 
 /**
  * Class GetImage
@@ -23,6 +22,9 @@ class GetImage extends AbstractWebService implements WsInterface
 {
     use WsAstrobinTrait;
 
+    /**
+     * @var string
+     */
     final public const END_POINT = 'image';
 
     protected function getEndPoint(): string
@@ -125,6 +127,7 @@ class GetImage extends AbstractWebService implements WsInterface
      * Get image filtered by range date
      * @throws JsonException
      * @throws WsException
+     * @throws \Exception
      */
     public function getImagesByRangeDate(
         ?string $dateFromStr,
@@ -164,7 +167,7 @@ class GetImage extends AbstractWebService implements WsInterface
             return null;
         }
 
-        $params = array_filter($filters, static fn($key) => in_array($key, array_column(ImageFilters::cases(), 'value'), true), ARRAY_FILTER_USE_KEY);
+        $params = array_filter($filters, static fn($key): bool => in_array($key, array_column(ImageFilters::cases(), 'value'), true), ARRAY_FILTER_USE_KEY);
         $params = [...$params, QueryFilters::LIMIT->value => $limit];
 
         return $this->sendRequestAndBuildResponse(null, $params);
