@@ -23,7 +23,7 @@ class TodayTest extends TestCase
     public function testPropertiesExist(): void
     {
         $reflexion = new \ReflectionClass((new Today()));
-        $props = array_map(static fn(\ReflectionProperty $prop) => $prop->getName(), $reflexion->getProperties());
+        $props = array_map(static fn(\ReflectionProperty $prop): string => $prop->getName(), $reflexion->getProperties());
         sort($props);
         $this->assertEquals(self::$expectedProperties, $props);
     }
