@@ -15,9 +15,10 @@ class GetUserTest extends TestCase
 {
 
     public ?GetUser $astrobinWs = null;
+
     public ?GetUser $badAstrobinWs = null;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         $this->badAstrobinWs = new GetUser(null, null);
         $astrobinKey = getenv('ASTROBIN_API_KEY');
@@ -30,12 +31,12 @@ class GetUserTest extends TestCase
         $reflection = new \ReflectionClass($this->astrobinWs);
         $method = $reflection->getMethod('getEndPoint');
         $method->setAccessible(true);
+
         $endPoint = $method->invoke($this->astrobinWs);
         $this->assertEquals(GetUser::END_POINT, $endPoint);
     }
 
     /**
-     * @return void
      * @throws \ReflectionException
      */
     public function testGetObjectEntity(): void
@@ -43,6 +44,7 @@ class GetUserTest extends TestCase
         $reflection = new \ReflectionClass($this->astrobinWs);
         $method = $reflection->getMethod('getObjectEntity');
         $method->setAccessible(true);
+
         $response = $method->invoke($this->astrobinWs);
         $this->assertEquals(User::class, $response);
     }
@@ -94,7 +96,7 @@ class GetUserTest extends TestCase
          * Test with limit too high
          */
         $user = 'siovene';
-        $response = $this->astrobinWs->getByUsername($user, 999999999);
+        $response = $this->astrobinWs->getByUsername($user, 999_999_999);
         $this->assertNull($response);
 
         /**
@@ -106,7 +108,7 @@ class GetUserTest extends TestCase
         $this->assertEquals($user, $response->username);
     }
 
-    public function tearDown(): void
+    protected function tearDown(): void
     {
         $this->astrobinWs = null;
         $this->badAstrobinWs = null;
